feat: add weekly and monthly aggregate materialized views (#457)
* feat: add weekly and monthly aggregate materialized views

* This will create *new* views called
  Events(Daily/Weekly/Monthly)To(Artifact/Project)
* Note that the old materialized views: EventsDailyByArtifact and
  EventsDailyByProject will still be around and need to be cleaned up in
  a future migration
* This only deals with the migrations to create new views. We will be
  populating the views manually and the frontend still needs to be wired
  up

* fixed tests

* nit

---------

+++ indexer/src/db/testing.ts
@@ -17,6 +17,9 @@ export async function clearDb() {
   const c = await initializeOnce();
   const entities = c.entityMetadatas;
   for (const entity of entities) {
+    if (entity.tableMetadataArgs.type === "view") {
+      continue;
+    }
     const repository = AppDataSource.getRepository(entity.name);
     await repository.query(
       `TRUNCATE ${entity.tableName} RESTART IDENTITY CASCADE`,
